requirements.txt
google-api-core==2.19.0
google-auth==2.29.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
protobuf==4.25.3
packaging==24.0
python-dotenv==1.0.1
pytz==2024.1
rsa==4.9
tqdm==4.66.4
typing-extensions==4.12.2
